Make m2e-android work with Andmore
This makes m2e-android use Andmore instead of ADT as part
of the import process.  Easiest way to test this is to setup an eclipse
environment with Andmore plugins in the works space.  Then import
the m2e-android andmore branch into the work space.  The only plugin
you need to bring is the main me.gladwell.eclipse.m2e.android plugin
everything.

Once these are in the workspace, run the andmore application, and import
an existing android maven project.